refactor: move processSprinklers to its own file
