+++ src/game-logic/reducers/index.js
@@ -99,7 +99,8 @@ import { decrementItemFromInventory } from './decrementItemFromInventory'
 import { applyCrows } from './applyCrows'
 import { showNotification } from './showNotification'
 import { processWeather } from './processWeather'
-import { modifyFieldPlotAt, processField, processSprinklers } from './field'
+import { modifyFieldPlotAt, processField } from './field'
+import { processSprinklers } from './processSprinklers'
 
 export * from './field'
 export * from './showNotification'
@@ -112,6 +113,7 @@ export * from './applyCrows'
 export * from './waterField'
 export * from './waterAllPlots'
 export * from './processWeather'
+export * from './processSprinklers'
 
 ///////////////////////////////////////////////////////////
 //
